fix(video-on-demand): video does not play in Chrome for Android
Apparently Chrome for Android has native support for HLS. However, for some reason it does not like our stream. After hours of debugging, best I can tell is that there is an HTTP header somewhere it does not like.

The workaround is to always use hls.js if it is supported by the browser. (Safari on iOS does not have support for Media Source Extensions (MSE). Older iOS does not support Managed Media Source (MMS). So older iOS cannot use hls.js.) The downside to this is that we seem to break AirPlay on all but (presumably) older iOS due to what seems to be an iOS bug with MMS: https://github.com/video-dev/hls.js/issues/6482#issuecomment-2159399478